refactor error handling to use external function
